Scripts, waypointArrow and car stops
Add waypointArrow scripts for the HUD display scene and the body locked display scene. The waypointArrow will point at the AICar to guide the participant in knowing where the AICar is coming from. Also, added car stops to all of the scenes to counteract learning. The total number of scenes is now 14.