WIP
Changed block order from fully random to balanced latin squares method. Small fix to HMD tracking as well, changed from rotating complete participant to just the camera on the place of the eyes. I removed the head of the prefab else this way in the way. This resolves unwanted looking around the car if you tilt your head / the HMD sideways ("roll"). Not perfect, but better.